requirements.txt
numpy
pandas
python-dotenv
pymysql
scikit-learn
# -e .